move ConsensusType
