+++ consensus/src/main/java/org/apache/iotdb/consensus/common/ConsensusType.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.confignode.consensus;
+package org.apache.iotdb.consensus.common;
 
 import java.util.Arrays;
